moved routines and activities initial states to App file
