+++ src/components/Routines.js
@@ -1,18 +1,4 @@
-import React, { useEffect, useState } from "react";
-import { callApi } from "../api";
-
-const Routines = () => {
-  const [routines, setRoutines] = useState([]);
-
-  const fetchRoutines = async () => {
-    const data = await callApi({ url: "/routines" });
-    setRoutines(data);
-  };
-
-  useEffect(() => {
-    fetchRoutines();
-  }, []);
-
+const Routines = ({ routines }) => {
   return (
     <div>
       <h1>Routines Page</h1>
